#24 - moved Payments integration tests to PaymentsIntegrationTest
